backend/requirements.txt
flask
pytesseract
pillow
opencv-python
numpy
pytest